requirements.txt
pydantic>=2.0,<3.0
PyYAML>=6.0
anthropic>=0.40.0
requests>=2.31.0
